Clarifying comment re: "if matrix unchanged" issue
The discussion boards are full with people wondering about the portion of the assignment worded "If the inverse has already been calculated *(and the matrix has not changed)*, then the cachesolve should retrieve the inverse from the cache."

A truly defensive implementation would pass the matrix into the getInverse function and check to see whether the passed-in matrix is truly the same as the matrix from which the cached inverse was calculated.

However, for clarity's sake I've decided not to implement this way.  Added comments to clarify that.